The Milo constructor

The Milo class extends the SingleCellExperiment class and is designed to work with neighbourhoods of cells. Therefore, it inherits from the SingleCellExperiment class and follows the same usage conventions. There is additional support for cell-to-cell distances via distance, and the KNN-graph used to define the neighbourhoods.

Milo(
  ...,
  graph = list(),
  nhoodDistances = Matrix(0L, sparse = TRUE),
  nhoods = list(),
  nhoodCounts = Matrix(0L, sparse = TRUE),
  nhoodIndex = list(),
  nhoodExpression = Matrix(0L, sparse = TRUE)
)

Arguments

...
Arguments passed to the Milo constructor to fill the slots of the base class. This should be either a SingleCellExperiment or matrix of features X cells

graph
An igraph object or list of adjacent vertices that represents the KNN-graph

nhoodDistances
A list containing sparse matrices of cell-to-cell distances for cells in the same neighbourhoods, one list entry per neighbourhood.

nhoods
A list of graph vertices, each containing the indices of the constiuent graph vertices in the respective neighbourhood

nhoodCounts
A matrix of neighbourhood X sample counts of the number of cells in each neighbourhood derived from the respective samples

nhoodIndex
A list of cells that are the neighborhood index cells.

nhoodExpression
A matrix of gene X neighbourhood expression.

Value

a Milo object

Details

In this class the underlying structure is the gene/feature X cell expression data. The additional slots provide a link between these single cells and the neighbourhood representation. This can be further extended by the use of an abstracted graph for visualisation that preserves the structure of the single-cell KNN-graph

A Milo object can also be constructed by inputting a feature X cell gene expression matrix. In this case it simply constructs a SingleCellExperiment and fills the relevant slots, such as reducedDims.

Examples

library(SingleCellExperiment)

#> Loading required package: SummarizedExperiment

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following object is masked from ‘package:limma’:
#> 
#>     plotMA

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> Loading required package: BiocParallel

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

#> 
#> Attaching package: ‘SingleCellExperiment’

#> The following object is masked from ‘package:edgeR’:
#> 
#>     cpm

ux <- matrix(rpois(12000, 5), ncol=200)
vx <- log2(ux + 1)
pca <- prcomp(t(vx))

sce <- SingleCellExperiment(assays=list(counts=ux, logcounts=vx),
                            reducedDims=SimpleList(PCA=pca$x))

milo <- Milo(sce)
milo

#> class: Milo 
#> dim: 60 200 
#> metadata(0):
#> assays(2): counts logcounts
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):
#> reducedDimNames(1): PCA
#> spikeNames(0):
#> altExpNames(0):
#> nhoods dimensions(1): 0
#> nhoodCounts dimensions(2): 1 1
#> nhoodDistances dimension(1): 0
#> graph names(0):
#> nhoodIndex names(1): 0
#> nhoodExpression dimension(2): 1 1
#> nhoodReducedDim names(0):
#> nhoodGraph names(0):
#> nhoodAdjacency dimension(0):