Build an abstracted graph of neighbourhoods for visualization

Build an abstracted graph of neighbourhoods for visualization

buildNhoodGraph(x, overlap = 1)

Arguments

x
A Milo object with a non-empty nhoods slot.

overlap
A numeric scalar that thresholds graph edges based on the number of overlapping cells between neighbourhoods.

Value

A Milo object with filled

Details

This constructs a graph where nodes represent neighbourhoods and edges represent the number of overlapping cells between two neighbourhoods.

Examples

NULL

#> NULL