The Milo container class

The Milo container class

Slots

graph

An igraph object that represents the kNN graph

nhoods

A list of neighbourhoods as graph indices and their constituent single cells

nhoodDistances

An list of PxN sparse matrices of Euclidean distances between vertices in each neighbourhood, one matrix per neighbourhood

nhoodCounts

An NxM sparse matrix of cells counts in each neighourhood across M samples

nhoodIndex

A list of the index vertices for each neighbourhood

nhoodExpression

An GxN matrix of genes X neighbourhoods containing average gene expression levels across cells in each neighbourhood

nhoodReducedDim

a list of reduced dimensional representations of neighbourhoods, including projections into lower dimension space

nhoodGraph

an igraph object that represents the graph of neighbourhoods