Perform post-hoc differential gene expression analysis

This function will perform differential gene expression analysis with differentially abundant neighbourhoods, by first aggregating adjacent and concordantly DA neighbourhoods, then comparing cells within these aggregated groups for differential gene experession using the input design. For comparing between DA neighbourhoods see findNhoodMarkers.

Arguments

x
A Milo object containing single-cell gene expression and neighbourhoods.

da.res
A data.frame containing DA results, as expected from running testNhoods.

da.fdr
A numeric scalar that determines at what FDR neighbourhoods are declared DA for the purposes of aggregating across concorantly DA neighbourhoods.

design
A formula or model.matrix object describing the experimental design for differential gene expression testing. The last component of the formula or last column of the model matrix are by default the test variable. This behaviour can be overridden by setting the model.contrasts argument. This should be the same as was used for DA testing.

meta.data
A cell X variable data.frame containing single-cell meta-data to which design refers. The order of rows (cells) must be the same as the Milo object columns.

model.contrasts
A string vector that defines the contrasts used to perform DA testing. This should be the same as was used for DA testing.

assay
A character scalar determining which assays slot to extract from the Milo object to use for DGE testing.

overlap
A scalar integer that determines the number of cells that must overlap between adjacent neighbourhoods for merging.

lfc.threshold
A scalar that determines the absolute log fold change above which neighbourhoods should be considerd 'DA' for merging. Default=NULL

merge.discord
A logical scalar that overrides the default behaviour and allows adjacent neighbourhoods to be merged if they have discordant log fold change signs. Using this argument is generally discouraged, but may be useful for constructing an empirical null group of cells, regardless of DA sign.

subset.row
A logical, integer or character vector indicating the rows of x to use for sumamrizing over cells in neighbourhoods.

gene.offset
A logical scalar the determines whether a per-cell offset is provided in the DGE GLM to adjust for the number of detected genes with expression > 0.

n.coef
A numeric scalar refering to the coefficient to select from the DGE model. This is especially pertinent when passing an ordered variable and only one specific type of effects are to be tested.

na.function
A valid NA action function to apply, should be one of na.fail, na.omit, na.exclude, na.pass.

compute.new
A logical scalar indicating whether to force computing a new neighbourhood adjacency matrix if already present.

Value

A list containing a data.frame of DGE results for each aggregated group of neighbourhoods.

Details

Adjacent neighbourhoods are first merged based on two criteria: 1) they share at least overlap number of cells, and 2) the DA log fold change sign is concordant. This behaviour can be modulated by setting overlap to be more or less stringent. Additionally, a threshold on the log fold-changes can be set, such that lfc.threshold is required to merge adjacent neighbourhoods. Note: adjacent neighbourhoods will never be merged with opposite signs unless merge.discord=TRUE.

Within each aggregated group of cells differential gene expression testing is performed using the single-cell log normalized gene expression with a GLM (for details see limma-package), or the single-cell counts using a negative binomial GLM (for details see edgeR-package). When using single-cell data for DGE it is recommended to set gene.offset=TRUE as this behaviour adjusts the model by the number of detected genes in each cell as a proxy for differences in capture efficiency and cellular RNA content.

Examples

data(sim_discrete)

milo <- Milo(sim_discrete$SCE)
milo <- buildGraph(milo, k=20, d=10, transposed=TRUE)

#> Constructing kNN graph with k:20

milo <- makeNhoods(milo, k=20, d=10, prop=0.3)

#> Checking valid object

meta.df <- sim_discrete$meta
meta.df$SampID <- paste(meta.df$Condition, meta.df$Replicate, sep="_")
milo <- countCells(milo, meta.data=meta.df, samples="SampID")

#> Checking meta.data validity

#> Setting up matrix with 118 neighbourhoods

#> Counting cells in neighbourhoods

test.meta <- data.frame("Condition"=c(rep("A", 3), rep("B", 3)), "Replicate"=rep(c("R1", "R2", "R3"), 2))
test.meta$Sample <- paste(test.meta$Condition, test.meta$Replicate, sep="_")
rownames(test.meta) <- test.meta$Sample
da.res <- testNhoods(milo, design=~Condition, design.df=test.meta[colnames(nhoodCounts(milo)), ])

#> Performing spatial FDR correction withk-distance weightingPerforming spatial FDR correction withneighbour-distance weightingPerforming spatial FDR correction withedge weightingPerforming spatial FDR correction withvertex weightingPerforming spatial FDR correction withnone weighting

nhood.dge <- testDiffExp(milo, da.res, da.fdr=0.2, design=~Condition, meta.data=meta.df, overlap=1, compute.new=TRUE)

#> Found 91 DA neighbourhoods at FDR 20%

#> Calculating nhood adjacency

#> Nhoods aggregated into 3 groups

nhood.dge

#> $`1`
#>                 logFC  AveExpr           t     P.Value adj.P.Val         B
#> Gene93   1.2318879299 4.168915  2.80047200 0.005263073 0.3726214 -4.409388
#> Gene67   0.8344175102 4.552016  2.50716264 0.012427094 0.3726214 -4.451473
#> Gene88   1.2551585224 3.867598  2.49131038 0.012989672 0.3726214 -4.453620
#> Gene2    0.9390474896 4.378169  2.44155490 0.014904857 0.3726214 -4.460276
#> Gene43   1.0142479922 4.115971  2.28132695 0.022869963 0.4573993 -4.480828
#> Gene29   0.9538862785 4.172319  2.19110302 0.028818643 0.4742882 -4.491807
#> Gene51   0.6986790876 4.508200  2.04989482 0.040798057 0.4742882 -4.508126
#> Gene56   0.9164785058 4.075053  2.03724208 0.042053243 0.4742882 -4.509537
#> Gene44   0.7882081040 4.299106  1.96000468 0.050446236 0.4742882 -4.517964
#> Gene89   0.9062371088 4.017089  1.91634779 0.055783882 0.4742882 -4.522586
#> Gene50   0.8849748634 4.037724  1.89644534 0.058369373 0.4742882 -4.524660
#> Gene21   0.8225050627 4.154201  1.87615496 0.061107217 0.4742882 -4.526752
#> Gene82   0.6710080463 4.458411  1.87216845 0.061657471 0.4742882 -4.527160
#> Gene31   0.9161185364 3.870317  1.82006273 0.069235227 0.4750520 -4.532421
#> Gene8    0.9222018015 3.814897  1.80539226 0.071502047 0.4750520 -4.533876
#> Gene19   0.7436454803 4.229951  1.77732761 0.076008312 0.4750520 -4.536627
#> Gene1    0.8277092082 3.963399  1.72261933 0.085459928 0.5027055 -4.541869
#> Gene84   0.7795768797 3.954160  1.62765543 0.104109956 0.5387679 -4.550585
#> Gene74   0.6252095868 4.344910  1.55351173 0.120815174 0.5387679 -4.557053
#> Gene26   0.6049193963 4.348093  1.54907550 0.121877806 0.5387679 -4.557430
#> Gene94   0.6228816931 4.308358  1.53900621 0.124316924 0.5387679 -4.558283
#> Gene57   0.5932341680 4.306711  1.46535246 0.143335638 0.5387679 -4.564356
#> Gene73   0.5840537784 4.329540  1.46461704 0.143536232 0.5387679 -4.564415
#> Gene25   0.5976308327 4.278826  1.44079664 0.150150874 0.5387679 -4.566315
#> Gene46   0.5168571650 4.443073  1.43689782 0.151255376 0.5387679 -4.566624
#> Gene23   0.5734601625 4.307268  1.41399250 0.157869931 0.5387679 -4.568417
#> Gene38   0.5497577960 4.358173  1.41336198 0.158055069 0.5387679 -4.568466
#> Gene7    0.5794504987 4.268111  1.40003812 0.162005978 0.5387679 -4.569496
#> Gene79   0.7968506608 3.567954  1.39110219 0.164697308 0.5387679 -4.570181
#> Gene35   0.6984887515 3.844366  1.38072339 0.167865409 0.5387679 -4.570971
#> Gene90   0.4746485446 4.494207  1.36801581 0.171806569 0.5387679 -4.571931
#> Gene39   0.4465814152 4.586457  1.36610317 0.172405721 0.5387679 -4.572074
#> Gene65  -0.5488523643 4.194656 -1.27866850 0.201495948 0.6105938 -4.578431
#> Gene91   0.4116072619 4.534313  1.23696388 0.216572444 0.6249532 -4.581316
#> Gene60   0.4170198890 4.523731  1.23099092 0.218796653 0.6249532 -4.581722
#> Gene59   0.5188118754 4.200303  1.20714664 0.227839584 0.6249532 -4.583321
#> Gene83   0.6478678225 3.691063  1.19837448 0.231232687 0.6249532 -4.583902
#> Gene100  0.4025223999 4.459586  1.10772823 0.268412090 0.6794166 -4.589655
#> Gene3    0.4383963155 4.281526  1.08927083 0.276460956 0.6794166 -4.590772
#> Gene33   0.4164993667 4.364404  1.08431975 0.278647789 0.6794166 -4.591068
#> Gene9   -0.4425494680 4.208725 -1.04163475 0.297990206 0.6794166 -4.593568
#> Gene99   0.4661657294 4.102530  1.02849468 0.304121375 0.6794166 -4.594317
#> Gene53   0.4133109098 4.299302  1.01351041 0.311214848 0.6794166 -4.595160
#> Gene17   0.3947261950 4.343804  1.00251626 0.316488454 0.6794166 -4.595771
#> Gene58  -0.4273885516 4.189548 -0.98735475 0.323856904 0.6794166 -4.596602
#> Gene20   0.4318138901 4.145323  0.98421318 0.325397601 0.6794166 -4.596772
#> Gene75   0.3583116370 4.442805  0.98277998 0.326102059 0.6794166 -4.596850
#> Gene40   0.3844237248 4.331774  0.97045848 0.332199385 0.6794166 -4.597513
#> Gene41   0.4142249452 4.162412  0.94746322 0.343774783 0.6794166 -4.598729
#> Gene95   0.4206180358 4.010282  0.89864091 0.369195527 0.6794166 -4.601213
#> Gene16   0.3006064862 4.519260  0.87287175 0.383073535 0.6794166 -4.602472
#> Gene13  -0.3920948925 4.077743 -0.86636716 0.386626592 0.6794166 -4.602784
#> Gene32   0.3653657373 4.215834  0.86463080 0.387578459 0.6794166 -4.602867
#> Gene87   0.4810596758 3.593742  0.84873364 0.396359632 0.6794166 -4.603618
#> Gene34   0.3275382688 4.363628  0.84531041 0.398266174 0.6794166 -4.603778
#> Gene5    0.3440518612 4.330202  0.84491874 0.398484661 0.6794166 -4.603796
#> Gene10   0.3583888530 4.141637  0.81944245 0.412851349 0.6794166 -4.604966
#> Gene61   0.3456189031 4.211983  0.81885500 0.413186218 0.6794166 -4.604992
#> Gene12   0.2781900253 4.507484  0.81597814 0.414828445 0.6794166 -4.605122
#> Gene76   0.3473011142 4.163919  0.80229050 0.422694731 0.6794166 -4.605733
#> Gene71   0.2569512480 4.565016  0.79271590 0.428248943 0.6794166 -4.606154
#> Gene66   0.3378790468 4.183129  0.78706067 0.431549438 0.6794166 -4.606401
#> Gene42   0.3172076610 4.294789  0.78611810 0.432100971 0.6794166 -4.606441
#> Gene47   0.3734236578 3.977207  0.78147021 0.434826614 0.6794166 -4.606642
#> Gene86   0.2949105357 4.224372  0.69569232 0.486884239 0.7378205 -4.610139
#> Gene78   0.2683763532 4.347768  0.69556889 0.486961507 0.7378205 -4.610143
#> Gene18   0.2349516394 4.446995  0.63560155 0.525272506 0.7706049 -4.612347
#> Gene98   0.2876529130 4.076740  0.62870833 0.529773367 0.7706049 -4.612587
#> Gene14  -0.2737685958 4.122243 -0.61720350 0.537328863 0.7706049 -4.612983
#> Gene69   0.2600895944 4.200389  0.60695292 0.544106079 0.7706049 -4.613329
#> Gene36   0.2299072747 4.380551  0.59705159 0.550692592 0.7706049 -4.613659
#> Gene70   0.2244171351 4.392637  0.59085352 0.554835563 0.7706049 -4.613862
#> Gene97   0.2278320614 4.342836  0.57874652 0.562972034 0.7711946 -4.614253
#> Gene15  -0.2769318965 3.662157 -0.50089911 0.616621287 0.8318784 -4.616574
#> Gene77  -0.2148992524 4.124432 -0.48118970 0.630552774 0.8318784 -4.617108
#> Gene85   0.1550213587 4.533009  0.45767407 0.647348205 0.8318784 -4.617718
#> Gene68   0.2233712392 3.886197  0.45196268 0.651455166 0.8318784 -4.617862
#> Gene49   0.1576755505 4.475648  0.44649492 0.655396894 0.8318784 -4.617997
#> Gene28  -0.2195382376 3.868209 -0.43559252 0.663285180 0.8318784 -4.618263
#> Gene72   0.1654583707 4.375627  0.43253704 0.665502713 0.8318784 -4.618336
#> Gene27   0.1630695767 4.312371  0.40150172 0.688190194 0.8496175 -4.619051
#> Gene52   0.1398489690 4.411239  0.37300768 0.709271204 0.8649649 -4.619661
#> Gene30   0.1235389283 4.507729  0.35829317 0.720246830 0.8662775 -4.619958
#> Gene4    0.1346422722 4.354220  0.33705569 0.736189870 0.8662775 -4.620366
#> Gene81   0.1579200787 4.027153  0.33686186 0.736335914 0.8662775 -4.620370
#> Gene96  -0.1184775428 4.350381 -0.30295001 0.762030367 0.8860818 -4.620969
#> Gene92  -0.1172047992 4.252990 -0.28075012 0.778996449 0.8953982 -4.621328
#> Gene64   0.1078928282 4.263859  0.26363047 0.792152950 0.8993234 -4.621585
#> Gene22  -0.1048115066 4.277137 -0.25294160 0.800397838 0.8993234 -4.621738
#> Gene55   0.1078198370 3.931902  0.22050995 0.825547178 0.9172746 -4.622162
#> Gene62   0.0681228496 4.494814  0.19697558 0.843911706 0.9273755 -4.622434
#> Gene54   0.0687031661 4.256143  0.16515137 0.868879095 0.9397551 -4.622752
#> Gene80   0.0656070265 4.248436  0.15868110 0.873972202 0.9397551 -4.622810
#> Gene11  -0.0470054170 4.056020 -0.10198788 0.918799523 0.9774463 -4.623218
#> Gene63   0.0306093013 4.463769  0.08389909 0.933163953 0.9822778 -4.623311
#> Gene37   0.0178695534 4.230325  0.04178775 0.966681459 0.9983606 -4.623458
#> Gene6    0.0183743255 3.890250  0.03673779 0.970705994 0.9983606 -4.623469
#> Gene45   0.0092123591 3.868750  0.01818328 0.985498538 0.9983606 -4.623497
#> Gene24  -0.0046759181 4.311708 -0.01148780 0.990837989 0.9983606 -4.623502
#> Gene48  -0.0008627031 4.269449 -0.00205552 0.998360600 0.9983606 -4.623506
#>         Nhood.Group
#> Gene93            1
#> Gene67            1
#> Gene88            1
#> Gene2             1
#> Gene43            1
#> Gene29            1
#> Gene51            1
#> Gene56            1
#> Gene44            1
#> Gene89            1
#> Gene50            1
#> Gene21            1
#> Gene82            1
#> Gene31            1
#> Gene8             1
#> Gene19            1
#> Gene1             1
#> Gene84            1
#> Gene74            1
#> Gene26            1
#> Gene94            1
#> Gene57            1
#> Gene73            1
#> Gene25            1
#> Gene46            1
#> Gene23            1
#> Gene38            1
#> Gene7             1
#> Gene79            1
#> Gene35            1
#> Gene90            1
#> Gene39            1
#> Gene65            1
#> Gene91            1
#> Gene60            1
#> Gene59            1
#> Gene83            1
#> Gene100           1
#> Gene3             1
#> Gene33            1
#> Gene9             1
#> Gene99            1
#> Gene53            1
#> Gene17            1
#> Gene58            1
#> Gene20            1
#> Gene75            1
#> Gene40            1
#> Gene41            1
#> Gene95            1
#> Gene16            1
#> Gene13            1
#> Gene32            1
#> Gene87            1
#> Gene34            1
#> Gene5             1
#> Gene10            1
#> Gene61            1
#> Gene12            1
#> Gene76            1
#> Gene71            1
#> Gene66            1
#> Gene42            1
#> Gene47            1
#> Gene86            1
#> Gene78            1
#> Gene18            1
#> Gene98            1
#> Gene14            1
#> Gene69            1
#> Gene36            1
#> Gene70            1
#> Gene97            1
#> Gene15            1
#> Gene77            1
#> Gene85            1
#> Gene68            1
#> Gene49            1
#> Gene28            1
#> Gene72            1
#> Gene27            1
#> Gene52            1
#> Gene30            1
#> Gene4             1
#> Gene81            1
#> Gene96            1
#> Gene92            1
#> Gene64            1
#> Gene22            1
#> Gene55            1
#> Gene62            1
#> Gene54            1
#> Gene80            1
#> Gene11            1
#> Gene63            1
#> Gene37            1
#> Gene6             1
#> Gene45            1
#> Gene24            1
#> Gene48            1
#> 
#> $`3`
#>              logFC  AveExpr         t      P.Value  adj.P.Val          B
#> Gene100 1.78455856 3.410554 3.5202226 0.0004376020 0.03695127 -0.2298877
#> Gene21  1.66334404 3.454423 3.3502233 0.0008174509 0.03695127 -0.7877573
#> Gene61  1.39855170 3.693066 3.2156767 0.0013151954 0.03695127 -1.2097136
#> Gene27  1.57780261 3.456953 3.1819034 0.0014780506 0.03695127 -1.3129161
#> Gene5   1.45383389 3.567069 3.1082978 0.0018993137 0.03798627 -1.5340606
#> Gene7   1.55503130 3.380371 3.0263832 0.0024960059 0.04160010 -1.7740842
#> Gene56  1.39714462 3.556930 2.9702555 0.0029991217 0.04284460 -1.9348473
#> Gene71  1.45096413 3.369591 2.8106961 0.0049755209 0.04389535 -2.3754247
#> Gene25  1.45438315 3.356594 2.8017575 0.0051151061 0.04389535 -2.3993867
#> Gene36  1.30274505 3.576823 2.8005733 0.0051338625 0.04389535 -2.4025556
#> Gene13  1.37590340 3.463492 2.7837790 0.0054066334 0.04389535 -2.4473511
#> Gene93  1.47848919 3.293818 2.7760728 0.0055361237 0.04389535 -2.4678157
#> Gene18  1.37752335 3.438188 2.7527105 0.0059459817 0.04389535 -2.5295101
#> Gene16  1.19105599 3.695110 2.7418672 0.0061453489 0.04389535 -2.5579677
#> Gene63  1.42519014 3.304101 2.6869927 0.0072496308 0.04833087 -2.7002595
#> Gene50  1.38975899 3.313870 2.6305009 0.0085695081 0.05001331 -2.8437402
#> Gene90  1.50815594 3.093909 2.6193329 0.0088545601 0.05001331 -2.8717442
#> Gene91  1.57490581 2.957614 2.5934470 0.0095481019 0.05001331 -2.9361956
#> Gene26  1.41237558 3.226715 2.5831828 0.0098362592 0.05001331 -2.9615746
#> Gene42  1.11503775 3.682624 2.5481531 0.0108789324 0.05001331 -3.0474299
#> Gene43  1.31351276 3.370969 2.5459441 0.0109478652 0.05001331 -3.0528048
#> Gene41  1.15473416 3.619901 2.5441884 0.0110029292 0.05001331 -3.0570734
#> Gene77  1.33371888 3.312375 2.5229079 0.0116901787 0.05082686 -3.1085773
#> Gene6   1.13609713 3.622809 2.5073326 0.0122170653 0.05090444 -3.1459993
#> Gene24  1.27973756 3.372075 2.4816615 0.0131315111 0.05252604 -3.2071721
#> Gene28  1.19444429 3.481628 2.4389406 0.0147879965 0.05307073 -3.3075781
#> Gene83  1.19164652 3.480051 2.4312625 0.0151045014 0.05307073 -3.3254388
#> Gene96  1.20111103 3.457035 2.4223406 0.0154797673 0.05307073 -3.3461224
#> Gene72  1.31096962 3.248837 2.4184326 0.0156467104 0.05307073 -3.3551582
#> Gene3   1.05452437 3.684196 2.4120857 0.0159212177 0.05307073 -3.3698021
#> Gene60  1.19782094 3.433602 2.3883696 0.0169848028 0.05478969 -3.4241804
#> Gene57  1.18348411 3.431604 2.3575437 0.0184601638 0.05715668 -3.4940577
#> Gene44  1.14777296 3.473393 2.3338215 0.0196707866 0.05715668 -3.5472141
#> Gene75  1.22964524 3.319004 2.3322806 0.0197517692 0.05715668 -3.5506483
#> Gene19  1.28630288 3.199130 2.3275007 0.0200048378 0.05715668 -3.5612870
#> Gene34  1.11398897 3.509512 2.3080073 0.0210664474 0.05733431 -3.6044474
#> Gene81  1.18364453 3.382135 2.3053715 0.0212136937 0.05733431 -3.6102553
#> Gene64  1.09969734 3.501733 2.2690522 0.0233358742 0.06141020 -3.6896103
#> Gene58  1.35607601 2.928418 2.2077348 0.0273381691 0.06967339 -3.8207248
#> Gene40  1.30397349 3.019870 2.2001905 0.0278693548 0.06967339 -3.8366086
#> Gene46  1.12593491 3.346642 2.1599804 0.0308531623 0.07525162 -3.9203500
#> Gene37  0.98086248 3.563035 2.0923617 0.0364900104 0.08688098 -4.0576895
#> Gene67  1.06190954 3.387001 2.0726934 0.0382860307 0.08883998 -4.0968176
#> Gene69  0.99299269 3.515731 2.0641468 0.0390895899 0.08883998 -4.1137050
#> Gene10  1.07209337 3.320082 2.0343397 0.0420051400 0.09296238 -4.1720553
#> Gene35  1.25525995 2.896726 2.0184152 0.0436368579 0.09296238 -4.2028812
#> Gene14  1.00073749 3.456680 2.0178829 0.0436923173 0.09296238 -4.2039075
#> Gene53  0.97846647 3.311637 1.8503500 0.0643613786 0.13300980 -4.5134452
#> Gene11  0.94464798 3.384754 1.8419942 0.0655746355 0.13300980 -4.5281815
#> Gene78  0.79966137 3.679459 1.8240508 0.0682438228 0.13300980 -4.5596013
#> Gene79  0.98770470 3.233694 1.8113830 0.0701815495 0.13300980 -4.5815980
#> Gene74  0.85101836 3.558329 1.8106265 0.0702986870 0.13300980 -4.5829068
#> Gene88  0.91429163 3.411590 1.8043884 0.0712706849 0.13300980 -4.5936779
#> Gene97  0.88640311 3.471744 1.8008603 0.0718252926 0.13300980 -4.5997534
#> Gene85  1.02947420 3.064679 1.7676911 0.0772141988 0.14038945 -4.6562898
#> Gene20  0.95023441 3.244587 1.7500627 0.0802097154 0.14323163 -4.6859092
#> Gene48  0.99677177 3.023657 1.6843395 0.0922200990 0.15871017 -4.7937222
#> Gene30  0.94176904 3.159388 1.6779261 0.0934656813 0.15871017 -4.8040217
#> Gene9   0.82487769 3.473130 1.6770392 0.0936390014 0.15871017 -4.8054431
#> Gene73  0.85162585 3.376927 1.6549410 0.0980410585 0.16340176 -4.8406119
#> Gene94  0.94570620 3.076780 1.6315417 0.1028810378 0.16493238 -4.8773433
#> Gene8   0.90918215 3.174209 1.6292418 0.1033668334 0.16493238 -4.8809253
#> Gene55  0.81403196 3.428716 1.6193672 0.1054733901 0.16493238 -4.8962473
#> Gene12  0.84444649 3.345166 1.6189798 0.1055567236 0.16493238 -4.8968465
#> Gene38  0.90657347 3.121850 1.5917660 0.1115426412 0.16920324 -4.9385810
#> Gene29  0.75813864 3.533561 1.5911813 0.1116741386 0.16920324 -4.9394700
#> Gene86  0.87830744 3.178053 1.5762717 0.1150686338 0.17174423 -4.9620267
#> Gene15  0.96529803 2.895833 1.5516240 0.1208576935 0.17773190 -4.9988506
#> Gene92  0.88678454 3.098767 1.5430695 0.1229193877 0.17814404 -5.0114955
#> Gene98  0.83557090 3.182115 1.5019464 0.1332161847 0.19030884 -5.0713057
#> Gene45  0.80759168 3.212761 1.4690644 0.1419201791 0.19778752 -5.1179678
#> Gene17  0.82527286 3.153976 1.4672716 0.1424070151 0.19778752 -5.1204822
#> Gene99  0.81031733 3.155209 1.4413741 0.1495833964 0.20302263 -5.1564612
#> Gene95  0.78098424 3.245859 1.4390639 0.1502367459 0.20302263 -5.1596396
#> Gene47  0.84524037 2.999329 1.4147697 0.1572398016 0.20827428 -5.1927551
#> Gene33  0.65875627 3.559706 1.4026438 0.1608265586 0.20827428 -5.2090731
#> Gene4   0.68472991 3.480512 1.3973536 0.1624106095 0.20827428 -5.2161482
#> Gene54  0.75932435 3.242287 1.3972094 0.1624539376 0.20827428 -5.2163406
#> Gene82  0.74092972 3.216081 1.3495425 0.1772646089 0.22438558 -5.2788776
#> Gene84  0.69616429 3.317274 1.3194964 0.1871038985 0.23387987 -5.3171816
#> Gene59  0.73232825 3.170105 1.3102272 0.1902192868 0.23483863 -5.3288242
#> Gene32  0.79966439 2.905953 1.2904851 0.1969817229 0.24022161 -5.3533479
#> Gene49  0.81298114 2.829130 1.2731028 0.2030801947 0.24467493 -5.3746321
#> Gene76  0.59674194 3.481803 1.2185987 0.2230923524 0.26558613 -5.4394994
#> Gene89  0.62089861 3.343375 1.1895047 0.2343352597 0.27396367 -5.4729635
#> Gene66  0.60259940 3.406090 1.1862722 0.2356087545 0.27396367 -5.4766316
#> Gene39  0.62598416 3.266106 1.1626058 0.2450819475 0.28170339 -5.5031834
#> Gene23  0.65342465 3.121423 1.1470955 0.2514336154 0.28572002 -5.5202944
#> Gene62  0.61368181 3.198474 1.1101429 0.2670264409 0.30002971 -5.5601349
#> Gene80  0.58331028 3.153736 1.0369837 0.2998270364 0.33314115 -5.6351635
#> Gene51  0.48972158 3.209230 0.8896144 0.3737442950 0.41070802 -5.7707757
#> Gene52  0.48223647 2.825016 0.7539513 0.4509376068 0.49014957 -5.8772756
#> Gene70  0.35938812 3.040827 0.6113786 0.5409952997 0.57987384 -5.9702525
#> Gene2   0.29553228 3.485363 0.6046084 0.5454849772 0.57987384 -5.9741847
#> Gene31  0.40011996 2.703610 0.5965092 0.5508801437 0.57987384 -5.9788312
#> Gene65  0.29416308 3.256645 0.5443164 0.5862642036 0.61069188 -6.0072712
#> Gene1   0.18530799 3.395486 0.3630582 0.7165869979 0.73874948 -6.0858282
#> Gene87  0.15259744 3.643330 0.3408099 0.7332705955 0.74374209 -6.0933080
#> Gene68  0.18983774 3.148381 0.3367822 0.7363046731 0.74374209 -6.0946116
#> Gene22  0.09088944 3.105680 0.1585812 0.8740095099 0.87400951 -6.1367749
#>         Nhood.Group
#> Gene100           3
#> Gene21            3
#> Gene61            3
#> Gene27            3
#> Gene5             3
#> Gene7             3
#> Gene56            3
#> Gene71            3
#> Gene25            3
#> Gene36            3
#> Gene13            3
#> Gene93            3
#> Gene18            3
#> Gene16            3
#> Gene63            3
#> Gene50            3
#> Gene90            3
#> Gene91            3
#> Gene26            3
#> Gene42            3
#> Gene43            3
#> Gene41            3
#> Gene77            3
#> Gene6             3
#> Gene24            3
#> Gene28            3
#> Gene83            3
#> Gene96            3
#> Gene72            3
#> Gene3             3
#> Gene60            3
#> Gene57            3
#> Gene44            3
#> Gene75            3
#> Gene19            3
#> Gene34            3
#> Gene81            3
#> Gene64            3
#> Gene58            3
#> Gene40            3
#> Gene46            3
#> Gene37            3
#> Gene67            3
#> Gene69            3
#> Gene10            3
#> Gene35            3
#> Gene14            3
#> Gene53            3
#> Gene11            3
#> Gene78            3
#> Gene79            3
#> Gene74            3
#> Gene88            3
#> Gene97            3
#> Gene85            3
#> Gene20            3
#> Gene48            3
#> Gene30            3
#> Gene9             3
#> Gene73            3
#> Gene94            3
#> Gene8             3
#> Gene55            3
#> Gene12            3
#> Gene38            3
#> Gene29            3
#> Gene86            3
#> Gene15            3
#> Gene92            3
#> Gene98            3
#> Gene45            3
#> Gene17            3
#> Gene99            3
#> Gene95            3
#> Gene47            3
#> Gene33            3
#> Gene4             3
#> Gene54            3
#> Gene82            3
#> Gene84            3
#> Gene59            3
#> Gene32            3
#> Gene49            3
#> Gene76            3
#> Gene89            3
#> Gene66            3
#> Gene39            3
#> Gene23            3
#> Gene62            3
#> Gene80            3
#> Gene51            3
#> Gene52            3
#> Gene70            3
#> Gene2             3
#> Gene31            3
#> Gene65            3
#> Gene1             3
#> Gene87            3
#> Gene68            3
#> Gene22            3
#> 
#> $`4`
#>            logFC  AveExpr        t      P.Value    adj.P.Val         B
#> Gene40  1.600685 3.623668 14.89978 3.864260e-50 3.864260e-48 103.08405
#> Gene99  1.578188 3.631507 14.73677 4.344994e-49 2.172497e-47 100.67339
#> Gene18  1.530893 3.663023 14.45631 2.626352e-47 8.754505e-46  96.58810
#> Gene32  1.519836 3.644528 14.26133 4.344609e-46 1.086152e-44  93.79409
#> Gene26  1.511058 3.647084 14.19191 1.169192e-45 2.338384e-44  92.80847
#> Gene95  1.550639 3.581577 14.14423 2.301200e-45 3.835333e-44  92.13436
#> Gene27  1.568582 3.553311 14.11326 3.568258e-45 5.097511e-44  91.69768
#> Gene44  1.508738 3.622197 14.03532 1.071652e-44 1.339565e-43  90.60297
#> Gene82  1.538909 3.569538 13.95123 3.486421e-44 3.873801e-43  89.42875
#> Gene74  1.500566 3.610308 13.88682 8.565890e-44 8.565890e-43  88.53407
#> Gene81  1.511206 3.591662 13.85751 1.287732e-43 1.170665e-42  88.12833
#> Gene87  1.524532 3.572262 13.83963 1.650729e-43 1.375608e-42  87.88118
#> Gene46  1.500971 3.601140 13.83015 1.882906e-43 1.448389e-42  87.75021
#> Gene21  1.520717 3.557378 13.70744 1.024839e-42 7.320276e-42  86.06416
#> Gene52  1.455131 3.650367 13.68240 1.445560e-42 9.092853e-42  85.72190
#> Gene94  1.466419 3.629437 13.68193 1.454857e-42 9.092853e-42  85.71552
#> Gene63  1.431241 3.694133 13.64775 2.324075e-42 1.367103e-41  85.24944
#> Gene29  1.449959 3.639705 13.58171 5.726579e-42 3.181433e-41  84.35218
#> Gene43  1.458137 3.624248 13.57615 6.177569e-42 3.251352e-41  84.27676
#> Gene38  1.480273 3.588763 13.55334 8.425221e-42 4.212610e-41  83.96803
#> Gene3   1.515595 3.531533 13.51499 1.418104e-41 6.752875e-41  83.45003
#> Gene69  1.496192 3.561078 13.50924 1.533063e-41 6.968467e-41  83.37248
#> Gene64  1.460922 3.605277 13.48818 2.038861e-41 8.864611e-41  83.08883
#> Gene47  1.442676 3.618336 13.39878 6.807699e-41 2.836541e-40  81.88947
#> Gene83  1.453763 3.600594 13.39156 7.501946e-41 3.000779e-40  81.79288
#> Gene88  1.423490 3.650627 13.38608 8.074098e-41 3.105422e-40  81.71977
#> Gene33  1.421179 3.649369 13.35850 1.169023e-40 4.329715e-40  81.35164
#> Gene61  1.484674 3.549928 13.33866 1.524849e-40 5.445889e-40  81.08733
#> Gene71  1.423663 3.629908 13.28549 3.102146e-40 1.069706e-39  80.38095
#> Gene98  1.431647 3.609194 13.24223 5.516853e-40 1.838951e-39  79.80837
#> Gene68  1.455462 3.575850 13.23671 5.936707e-40 1.915067e-39  79.73542
#> Gene37  1.444230 3.588405 13.22086 7.326868e-40 2.289646e-39  79.52618
#> Gene93  1.431250 3.605681 13.21676 7.735645e-40 2.344135e-39  79.47219
#> Gene20  1.437301 3.595866 13.20842 8.641054e-40 2.541487e-39  79.36211
#> Gene59  1.445525 3.575388 13.14323 2.045848e-39 5.732964e-39  78.50501
#> Gene80  1.434549 3.589909 13.14257 2.063867e-39 5.732964e-39  78.49629
#> Gene60  1.407560 3.628793 13.12940 2.455045e-39 6.635256e-39  78.32369
#> Gene36  1.410505 3.620715 13.11330 3.034846e-39 7.986436e-39  78.11287
#> Gene41  1.394839 3.647367 13.10168 3.535984e-39 9.066626e-39  77.96089
#> Gene12  1.442403 3.572369 13.09477 3.872266e-39 9.680665e-39  77.87056
#> Gene73  1.423224 3.597666 13.09103 4.067184e-39 9.919961e-39  77.82172
#> Gene28  1.430445 3.585204 13.07270 5.173763e-39 1.221550e-38  77.58243
#> Gene5   1.406080 3.620604 13.07155 5.252666e-39 1.221550e-38  77.56738
#> Gene86  1.441311 3.564851 13.03681 8.280791e-39 1.881998e-38  77.11476
#> Gene13  1.420944 3.590723 13.02345 9.860854e-39 2.191301e-38  76.94113
#> Gene96  1.440700 3.556780 12.98268 1.678950e-38 3.649892e-38  76.41200
#> Gene66  1.409222 3.598864 12.97007 1.978698e-38 4.209996e-38  76.24868
#> Gene34  1.428088 3.567583 12.93419 3.154419e-38 6.571707e-38  75.78502
#> Gene67  1.415408 3.583059 12.92074 3.756398e-38 7.642377e-38  75.61138
#> Gene24  1.431656 3.559885 12.91942 3.821188e-38 7.642377e-38  75.59438
#> Gene70  1.374016 3.647001 12.90442 4.641137e-38 9.100269e-38  75.40112
#> Gene9   1.420680 3.568777 12.87459 6.826987e-38 1.312882e-37  75.01747
#> Gene56  1.439853 3.537502 12.86963 7.278655e-38 1.373331e-37  74.95378
#> Gene6   1.396491 3.593225 12.81597 1.453907e-37 2.692421e-37  74.26600
#> Gene97  1.385405 3.608070 12.80781 1.614770e-37 2.891571e-37  74.16169
#> Gene42  1.378422 3.619150 12.80651 1.642109e-37 2.891571e-37  74.14500
#> Gene100 1.426350 3.548388 12.80622 1.648195e-37 2.891571e-37  74.14132
#> Gene45  1.372077 3.627701 12.79283 1.957643e-37 3.375246e-37  73.97029
#> Gene85  1.403466 3.579861 12.79033 2.021396e-37 3.426095e-37  73.93844
#> Gene57  1.434541 3.521788 12.74544 3.593862e-37 5.989770e-37  73.36647
#> Gene7   1.371811 3.614674 12.72027 4.958396e-37 8.128519e-37  73.04657
#> Gene92  1.404071 3.564235 12.69626 6.736173e-37 1.086479e-36  72.74203
#> Gene58  1.376980 3.593717 12.64013 1.375533e-36 2.183386e-36  72.03249
#> Gene53  1.355376 3.626710 12.63207 1.523724e-36 2.380818e-36  71.93081
#> Gene11  1.392726 3.561038 12.57481 3.145243e-36 4.838835e-36  71.21059
#> Gene75  1.387612 3.566309 12.55986 3.798069e-36 5.754649e-36  71.02318
#> Gene50  1.361118 3.599051 12.52850 5.638191e-36 8.415211e-36  70.63060
#> Gene76  1.393433 3.550271 12.52078 6.213498e-36 9.137497e-36  70.53406
#> Gene1   1.358533 3.599648 12.50844 7.255888e-36 1.051578e-35  70.37995
#> Gene55  1.380952 3.566801 12.50254 7.813527e-36 1.116218e-35  70.30638
#> Gene72  1.377706 3.564948 12.46207 1.297833e-35 1.827934e-35  69.80220
#> Gene39  1.334411 3.628466 12.44546 1.597482e-35 2.218725e-35  69.59580
#> Gene14  1.377577 3.560435 12.43456 1.830713e-35 2.476650e-35  69.46040
#> Gene51  1.393184 3.533827 12.43447 1.832721e-35 2.476650e-35  69.45931
#> Gene90  1.376040 3.559862 12.41740 2.267775e-35 3.014663e-35  69.24769
#> Gene35  1.335649 3.620564 12.41658 2.291144e-35 3.014663e-35  69.23751
#> Gene16  1.406094 3.500484 12.39948 2.835573e-35 3.682562e-35  69.02570
#> Gene89  1.336054 3.615468 12.39305 3.071732e-35 3.938117e-35  68.94622
#> Gene54  1.345076 3.594360 12.35135 5.157845e-35 6.528918e-35  68.43133
#> Gene49  1.353319 3.579662 12.33204 6.552773e-35 8.190966e-35  68.19354
#> Gene91  1.343141 3.592188 12.31975 7.629948e-35 9.419688e-35  68.04235
#> Gene2   1.370009 3.548632 12.30165 9.544023e-35 1.163905e-34  67.82000
#> Gene78  1.345976 3.585100 12.30008 9.731499e-35 1.172470e-34  67.80068
#> Gene22  1.342129 3.583204 12.25273 1.744956e-34 2.077328e-34  67.22062
#> Gene19  1.342112 3.574624 12.19820 3.409287e-34 4.010926e-34  66.55536
#> Gene77  1.368631 3.526746 12.18226 4.144582e-34 4.819281e-34  66.36138
#> Gene84  1.326272 3.587905 12.13785 7.130610e-34 8.196104e-34  65.82249
#> Gene30  1.338822 3.554898 12.05446 1.964961e-33 2.232910e-33  64.81584
#> Gene4   1.326244 3.571365 12.03417 2.511922e-33 2.822384e-33  64.57198
#> Gene25  1.323909 3.565923 11.98104 4.769609e-33 5.299566e-33  63.93527
#> Gene15  1.296314 3.599590 11.93523 8.271908e-33 9.090008e-33  63.38858
#> Gene8   1.299310 3.589008 11.89797 1.292407e-32 1.404790e-32  62.94556
#> Gene62  1.322602 3.543718 11.85149 2.251007e-32 2.420438e-32  62.39469
#> Gene10  1.273720 3.618368 11.82977 2.914848e-32 3.100902e-32  62.13813
#> Gene65  1.315096 3.521918 11.68478 1.617988e-31 1.703145e-31  60.43686
#> Gene48  1.271862 3.542902 11.39298 4.782006e-30 4.981256e-30  57.07675
#> Gene79  1.222841 3.589796 11.20235 4.174901e-29 4.304021e-29  54.92750
#> Gene23  1.265778 3.497723 11.15170 7.380068e-29 7.530681e-29  54.36253
#> Gene31  1.214042 3.584202 11.08919 1.485631e-28 1.500638e-28  53.66876
#> Gene17  1.263431 3.477887 11.05797 2.103747e-28 2.103747e-28  53.32383
#>         Nhood.Group
#> Gene40            4
#> Gene99            4
#> Gene18            4
#> Gene32            4
#> Gene26            4
#> Gene95            4
#> Gene27            4
#> Gene44            4
#> Gene82            4
#> Gene74            4
#> Gene81            4
#> Gene87            4
#> Gene46            4
#> Gene21            4
#> Gene52            4
#> Gene94            4
#> Gene63            4
#> Gene29            4
#> Gene43            4
#> Gene38            4
#> Gene3             4
#> Gene69            4
#> Gene64            4
#> Gene47            4
#> Gene83            4
#> Gene88            4
#> Gene33            4
#> Gene61            4
#> Gene71            4
#> Gene98            4
#> Gene68            4
#> Gene37            4
#> Gene93            4
#> Gene20            4
#> Gene59            4
#> Gene80            4
#> Gene60            4
#> Gene36            4
#> Gene41            4
#> Gene12            4
#> Gene73            4
#> Gene28            4
#> Gene5             4
#> Gene86            4
#> Gene13            4
#> Gene96            4
#> Gene66            4
#> Gene34            4
#> Gene67            4
#> Gene24            4
#> Gene70            4
#> Gene9             4
#> Gene56            4
#> Gene6             4
#> Gene97            4
#> Gene42            4
#> Gene100           4
#> Gene45            4
#> Gene85            4
#> Gene57            4
#> Gene7             4
#> Gene92            4
#> Gene58            4
#> Gene53            4
#> Gene11            4
#> Gene75            4
#> Gene50            4
#> Gene76            4
#> Gene1             4
#> Gene55            4
#> Gene72            4
#> Gene39            4
#> Gene14            4
#> Gene51            4
#> Gene90            4
#> Gene35            4
#> Gene16            4
#> Gene89            4
#> Gene54            4
#> Gene49            4
#> Gene91            4
#> Gene2             4
#> Gene78            4
#> Gene22            4
#> Gene19            4
#> Gene77            4
#> Gene84            4
#> Gene30            4
#> Gene4             4
#> Gene25            4
#> Gene15            4
#> Gene8             4
#> Gene62            4
#> Gene10            4
#> Gene65            4
#> Gene48            4
#> Gene79            4
#> Gene23            4
#> Gene31            4
#> Gene17            4
#>